Latest

+++ Assets/Scripts/Entity/EntityData.cs
@@ -3,10 +3,10 @@ using UnityEngine;
 public class EntityData : ScriptableObject
 {
     [Header("Entity Stats")]
-    public int strength;
-    public int defense;
-    public int speed;
-    public int health;
-    public int maxHealth;
+    public float strength;
+    public float defense;
+    public float speed;
+    public float health;
+    public float maxHealth;
     
 }
